+++ internal/templates/domains_service.go
@@ -1,6 +1,48 @@
 package templates
 
+import (
+	appdata "github.com/Drafteame/draft/internal/data"
+	"github.com/Drafteame/draft/internal/dtos"
+)
+
 type Service struct {
+	Postgres ServicePostgres
+	Dynamo   ServiceDynamo
+}
+
+func loadDomainsService(v *Service, data any) error {
+	input, ok := data.(dtos.DomainInput)
+	if !ok {
+		return nil
+	}
+
+	var loaders []func(*Service, any) error
+
+	switch input.DBType {
+	case appdata.DBTypePostgres:
+		loaders = append(loaders, loadServicePostgres)
+	case appdata.DBTypeDynamo:
+		loaders = append(loaders, loadServiceDynamo)
+	}
+
+	for _, loader := range loaders {
+		if err := loader(v, data); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func loadServicePostgres(v *Service, data any) error {
+	return loadDomainsServicePostgres(&v.Postgres, data)
+}
+
+func loadServiceDynamo(v *Service, data any) error {
+	return loadDomainsServiceDynamo(&v.Dynamo, data)
+}
+
+type ServicePostgres struct {
 	CreateGo        []byte
 	CreateTestGo    []byte
 	DeleteGo        []byte
@@ -16,25 +58,27 @@ type Service struct {
 	ServiceTestGo   []byte
 	UpdateGo        []byte
 	UpdateTestGo    []byte
+	ProvideGo       []byte
 }
 
-func loadDomainsService(v *Service, data any) error {
-	loaders := []func(*Service, any) error{
-		loadServiceCreateGo,
-		loadServiceCreateTestGo,
-		loadServiceDeleteGo,
-		loadServiceDeleteTestGo,
-		loadServiceGetGo,
-		loadServiceGetTestGo,
-		loadServiceInterfaces,
-		loadServiceSearchGo,
-		loadServiceSearchTestGo,
-		loadServiceSearchOneGo,
-		loadServiceSearchOneTestGo,
-		loadServiceServiceGo,
-		loadServiceServiceTestGo,
-		loadServiceUpdateGo,
-		loadServiceUpdateTestGo,
+func loadDomainsServicePostgres(v *ServicePostgres, data any) error {
+	loaders := []func(*ServicePostgres, any) error{
+		loadServicePostgresCreateGo,
+		loadServicePostgresCreateTestGo,
+		loadServicePostgresDeleteGo,
+		loadServicePostgresDeleteTestGo,
+		loadServicePostgresGetGo,
+		loadServicePostgresGetTestGo,
+		loadServicePostgresInterfacesGo,
+		loadServicePostgresSearchGo,
+		loadServicePostgresSearchTestGo,
+		loadServicePostgresSearchOneGo,
+		loadServicePostgresSearchOneTestGo,
+		loadServicePostgresServiceGo,
+		loadServicePostgresServiceTestGo,
+		loadServicePostgresUpdateGo,
+		loadServicePostgresUpdateTestGo,
+		loadServicePostgresProvideGo,
 	}
 
 	for _, loader := range loaders {
@@ -46,9 +90,9 @@ func loadDomainsService(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceCreateGo(v *Service, data any) error {
-	name := "domains/service/create.go"
-	path := "tmpl/domain/service/create.go.tmpl"
+func loadServicePostgresCreateGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/create.go"
+	path := "tmpl/domain/service/postgres/create.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -60,9 +104,9 @@ func loadServiceCreateGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceCreateTestGo(v *Service, data any) error {
-	name := "domains/service/create_test.go"
-	path := "tmpl/domain/service/create_test.go.tmpl"
+func loadServicePostgresCreateTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/create_test.go"
+	path := "tmpl/domain/service/postgres/create_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -74,9 +118,9 @@ func loadServiceCreateTestGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceDeleteGo(v *Service, data any) error {
-	name := "domains/service/delete.go"
-	path := "tmpl/domain/service/delete.go.tmpl"
+func loadServicePostgresDeleteGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/delete.go"
+	path := "tmpl/domain/service/postgres/delete.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -88,9 +132,9 @@ func loadServiceDeleteGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceDeleteTestGo(v *Service, data any) error {
-	name := "domains/service/delete_test.go"
-	path := "tmpl/domain/service/delete_test.go.tmpl"
+func loadServicePostgresDeleteTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/delete_test.go"
+	path := "tmpl/domain/service/postgres/delete_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -102,9 +146,9 @@ func loadServiceDeleteTestGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceGetGo(v *Service, data any) error {
-	name := "domains/service/get.go"
-	path := "tmpl/domain/service/get.go.tmpl"
+func loadServicePostgresGetGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/get.go"
+	path := "tmpl/domain/service/postgres/get.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -116,9 +160,9 @@ func loadServiceGetGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceGetTestGo(v *Service, data any) error {
-	name := "domains/service/get_test.go"
-	path := "tmpl/domain/service/get_test.go.tmpl"
+func loadServicePostgresGetTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/get_test.go"
+	path := "tmpl/domain/service/postgres/get_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -130,9 +174,9 @@ func loadServiceGetTestGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceInterfaces(v *Service, data any) error {
-	name := "domains/service/interfaces.go"
-	path := "tmpl/domain/service/interfaces.go.tmpl"
+func loadServicePostgresInterfacesGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/interfaces.go"
+	path := "tmpl/domain/service/postgres/interfaces.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -144,9 +188,9 @@ func loadServiceInterfaces(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceSearchGo(v *Service, data any) error {
-	name := "domains/service/search.go"
-	path := "tmpl/domain/service/search.go.tmpl"
+func loadServicePostgresSearchGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/search.go"
+	path := "tmpl/domain/service/postgres/search.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -158,9 +202,9 @@ func loadServiceSearchGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceSearchTestGo(v *Service, data any) error {
-	name := "domains/service/search_test.go"
-	path := "tmpl/domain/service/search_test.go.tmpl"
+func loadServicePostgresSearchTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/search_test.go"
+	path := "tmpl/domain/service/postgres/search_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -172,9 +216,9 @@ func loadServiceSearchTestGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceSearchOneGo(v *Service, data any) error {
-	name := "domains/service/search_one.go"
-	path := "tmpl/domain/service/search_one.go.tmpl"
+func loadServicePostgresSearchOneGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/search_one.go"
+	path := "tmpl/domain/service/postgres/search_one.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -186,9 +230,9 @@ func loadServiceSearchOneGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceSearchOneTestGo(v *Service, data any) error {
-	name := "domains/service/search_one_test.go"
-	path := "tmpl/domain/service/search_one_test.go.tmpl"
+func loadServicePostgresSearchOneTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/search_one_test.go"
+	path := "tmpl/domain/service/postgres/search_one_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -200,9 +244,9 @@ func loadServiceSearchOneTestGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceServiceGo(v *Service, data any) error {
-	name := "domains/service/service.go"
-	path := "tmpl/domain/service/service.go.tmpl"
+func loadServicePostgresServiceGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/service.go"
+	path := "tmpl/domain/service/postgres/service.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -214,9 +258,9 @@ func loadServiceServiceGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceServiceTestGo(v *Service, data any) error {
-	name := "domains/service/service_test.go"
-	path := "tmpl/domain/service/service_test.go.tmpl"
+func loadServicePostgresServiceTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/service_test.go"
+	path := "tmpl/domain/service/postgres/service_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -228,9 +272,9 @@ func loadServiceServiceTestGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceUpdateGo(v *Service, data any) error {
-	name := "domains/service/update.go"
-	path := "tmpl/domain/service/update.go.tmpl"
+func loadServicePostgresUpdateGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/update.go"
+	path := "tmpl/domain/service/postgres/update.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -242,9 +286,9 @@ func loadServiceUpdateGo(v *Service, data any) error {
 	return nil
 }
 
-func loadServiceUpdateTestGo(v *Service, data any) error {
-	name := "domains/service/update_test.go"
-	path := "tmpl/domain/service/update_test.go.tmpl"
+func loadServicePostgresUpdateTestGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/update_test.go"
+	path := "tmpl/domain/service/postgres/update_test.go.tmpl"
 
 	content, err := loadTemplate(name, path, data, domain)
 	if err != nil {
@@ -252,6 +296,84 @@ func loadServiceUpdateTestGo(v *Service, data any) error {
 	}
 
 	v.UpdateTestGo = content
+
+	return nil
+}
+
+func loadServicePostgresProvideGo(v *ServicePostgres, data any) error {
+	name := "domains/service/postgres/provide.go"
+	path := "tmpl/domain/service/postgres/provide.go.tmpl"
+
+	content, err := loadTemplate(name, path, data, domain)
+	if err != nil {
+		return err
+	}
+
+	v.ProvideGo = content
+
+	return nil
+}
+
+type ServiceDynamo struct {
+	InterfacesGo []byte
+	ServiceGo    []byte
+	ProviderGo   []byte
+}
+
+func loadDomainsServiceDynamo(v *ServiceDynamo, data any) error {
+	loaders := []func(*ServiceDynamo, any) error{
+		loadServiceDynamoInterfacesGo,
+		loadServiceDynamoServiceGo,
+		loadServiceDynamoProviderGo,
+	}
+
+	for _, loader := range loaders {
+		if err := loader(v, data); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func loadServiceDynamoInterfacesGo(v *ServiceDynamo, data any) error {
+	name := "domains/service/dynamo/interfaces.go"
+	path := "tmpl/domain/service/dynamo/interfaces.go.tmpl"
+
+	content, err := loadTemplate(name, path, data, domain)
+	if err != nil {
+		return err
+	}
+
+	v.InterfacesGo = content
+
+	return nil
+}
+
+func loadServiceDynamoServiceGo(v *ServiceDynamo, data any) error {
+	name := "domains/service/dynamo/service.go"
+	path := "tmpl/domain/service/dynamo/service.go.tmpl"
+
+	content, err := loadTemplate(name, path, data, domain)
+	if err != nil {
+		return err
+	}
+
+	v.ServiceGo = content
+
+	return nil
+}
+
+func loadServiceDynamoProviderGo(v *ServiceDynamo, data any) error {
+	name := "domains/service/dynamo/provider.go"
+	path := "tmpl/domain/service/dynamo/provider.go.tmpl"
+
+	content, err := loadTemplate(name, path, data, domain)
+	if err != nil {
+		return err
+	}
+
+	v.ProviderGo = content
 
 	return nil
 }